requirements.txt
asgiref==3.6.0
attrs==22.2.0
cfgv==3.3.1
distlib==0.3.6
Django==4.1.6
django-environ==0.9.0
django-filter==22.1
django-shortcuts==1.6
django-version==1.0.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
environ==1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.0.0
filelock==3.9.0
identify==2.5.17
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
Markdown==3.4.1
mysqlclient==2.1.1
nodeenv==1.7.0
packaging==23.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
pytest-is-running==1.3.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
virtualenv==20.19.0
